requirements.txt
fastapi~=0.115.12
uvicorn[standard]
python-multipart
pandas~=2.2.3
pytest
PyYAML~=6.0.2
pydantic~=2.11.5
httpx
